Add tmux to common system packages
When I do have to login to a machine, it's nice to have a terminal-
multiplexing capability and let me handle flakey VPN connections in
a graceful manner, so that any child processes aren't accidentally
killed.

Not sure if we should do it like this, or like we did for GOV.UK.

Thoughts?