Add admin.backdrop app
We need to have a suitable vhost in nginx. This adds that, plus gets us
closer to breaking out a separate admin application. Slightly nasty in
that we're going to be running 2 instances of write.backdrop until that
happens but hopefully the transition will be smoother when we do make
the leap.