Add kibana
We need this to be able to get a more holistic view over all of our
logs.

puppet-nginx module updated for custom log type support over all types
of vhost.

Made our own kibana class as all the ones out on the www are for kibana2
and that is old and dusty now.