Add local cname for www fqdn
Collectors and the smoke tests use the www fqdn to talk to services
inside of the infrastructure. Skyscape does not like us doing this and
has locked it down. So that these can still operate we have internal dns
to round robin route the fqdn to the frontend servers as the load
balancers would