Elasticsearch should store data in new mount
Add parameter to puppet-elasticsearch and switched to using the latest
commit in our Puppetfile (all the noise is around that change).

Moved defining elasticsearch class into monitoring class as I didn't
want to have the mounted directory hardcoded in two places.

Added a mount resource so that we can avoid trying to mount a disk that
isn't there when provisioning on vagrant (environment=development).