Fix nginx alerts so they are more reasonable
Previously, if there was no data in the period we were searching in
graphite it would cause a critical. This was less than ideal as no data
means we haven't had any 4xx or 5xx errors and should be taken as OK.

I made a change to our branch of sensu-community-plugins to add a flag
to ignore any no data errors and return an OK.

I have also changed fqdn to hostname as this is what logstash uses when
outputing to statsd.