Raise priority of environment (!?)
WARNING: I am not sure about this.

What is happening is that we are defining logstash::defaultsfiles in
both the role-monitoring file and in our environment specific files as
well. The result of this is that we will never take into account our
environment specific setting, as it will come later in the hierarchy.

Thus, I've switched it to take precedence over the roles. Let me know if
that is a terrible idea.