Add assets vhost to the frontend machines
Assets will be served from assets.env.pp-domain.

This commit adds a vhost listening on 80 and 443 on that subdomain to
the frontend machines, with a root inside /var/www/{subdomain} (by default).

It then adds some extra nginx config to the end of that server block
so that all requests under /spotlight are served from the public
directory of the current Spotlight deployment.

assets.perfplat.dev/spotlight/index.css -> /opt/spotlight/current/public/index.css

We set 2 hour cache times and send an etag. Once we've verified that we're doing
proper cachebusting in the app, we can increase that Expires header to months or
years.