Ignore too many processes
At the moment our alerts for mongo fire if there are more than one
process running. This causes regular firing when our sync scripts are
running. We are not bothered if there are too many process are running,
only too few so I have set the check to ignore too many processes.