We should add alerts for the rate of new logs
In order to clarify whether the logging pipeline is working we want to
check the rate of new logs entering elasticsearch. This works as we
expect a constant flow of logs to be picked up and elasticsearch is the
end of the line for our pipeline. If the rate hits zero then something
has broken in the pipeline