Add sensu checks for freshness of backups
We generate backups on the Mongo boxes and the PostgreSQL box with a
cron job. These are periodically copied over to the backups box by a
Jenkins job.

Here are sensu checks that the mongo and postgresql backup directories
on the backups box always contain files younger than 1 day old (ie that
they happen daily).