Add environment specific domain names
[fixes 68792636]
https://www.pivotaltracker.com/story/show/68792636

For admin and assets add environment specific domain names as cnames to
dnsmasq.

In production we serve www (for example) on
www.performance.service.gov.uk and on
www.production.performance.service.gov.uk. If one of these URLs is not
set up on dnsmasq it will resolve to our external IP address which isn't
accessible from within our infrastructure.