Switch nginx module to one maintained and modern
The motivation for making this change is as follows:
  - the nginx module we are using is a tweaked module forked from a very
    old nginx module,
  - we need to bring the management of vhost locations into the puppet
    code for the work we need to do to capture client side exceptions,
  - more flexibility on what can be configured through the module and a
    better than 'magic' way of configuring parameters that are not
    covered by the module

Used proxy_vhost as the interface as most of the vhosts were defined
using it. I've had to do a fair amount of translation between the old
paramters and how the new modules expect them.

Assets have been switched to use the new nginx location resources
directly as we can no longer directly inject config (which is good!).

I have dropped the nginx version changes as the version we were testing
has now been released into stable.