Fix expires directive for new nginx module
This was changed in ac5714d827bc782e78125ebe01efc6aeec42bcde when we
started to use the new module, but the docs for the module we're now
using say:

    location_custom_cfg - cannot be used with other location types (proxy, fastcgi, root, or stub_status)

Changing to location_cfg_prepend ensures that the expires directive
is present in the location block.

This will change the header from `Cache-Control: max-age=900` (default, set
by Varnish) to 2592000 (Spotlight) or 3600 (Stagecraft).