Change location of Spotlight assets on disk
We now have a shared directory and the deployment process
stores the app's assets in it on each deployment.

Change the place that the vhost looks for to serve up assets.

This will mean that we can continue to serve assets for old
deployed versions of the app, so that we should have no cases
where we 404 assets anymore.