Add SSL params to performanceplatform::app
We need to be able to set the ssl_cert/key for spotlight::app through
performanceplatform::app. Spotlight's vhost is always on an environment
domain (*.<env>.performance.service.gov.uk) so it needs to use the
environment_cert/key. By default it should fall back to using the
default of the proxy_vhost, which is public_cert/key.