Set xFilesFactor to zero for statsd timers
We have very low throughput on the performance platform which means that
timers often get aggregated to null because fewer than 10% of the values
are non-null. The risk with doing this change is that we will loose visibility
of flakeyness in metrics (ie. a metric stops sending for a few seconds,
it will still be aggregated to a number if there is any number in it).
That is why it is only being applied to statsd timers.