Live clone of postgres database
This introduces a script that can be run on a new secondary to bring it
up to speed with a primary without having to take the primary down.

On the priamry the checkpoint_segments and wal_keep_segments have been
increased to allow the secondary to catch up on writes that happen
during it copying the data over.

A login user has been introduced on the primary as well, as that is
needed for the pg_basebackup command.

This is modified slightly from
http://www.rassoc.com/gregr/weblog/2013/02/16/zero-to-postgresql-streaming-replication-in-10-mins/

I've gone with syncing to a separate directory and then rsyncing to
preserve things like the certificate and recovery.conf

This includes some fixes that @timmow noticed around rsync.