Upgrade postgresql version to 9.3
We need JSON fields that were introduced in version 9.2, as 9.3 has been
stable for over 6months we should move to that.

This includes the official apt repo from postgresql as this should be
preferred to the random PPA we had stolen the package from.

I haven't opted to use the 'manage_package_repo' parameter as it cannot
be defined through hiera
(https://tickets.puppetlabs.com/browse/MODULES-180) and trying to define
the postgresql::globals class and not ending up with duplicate
decoration was causing me to bang my head against the wall, so I just
ripped the code out of the module.