Add a request ID to all the things
As person responsible for operating software in production
I would like to have a unique ID associated with every request
So that it is easy for me to track down problems in a distributed
system.

This is the first step to that. Every request coming in to our SSL
termination machines will have a UUID assigned to it with an HTTP
header.

We can then amend our logging to include this header where appropriate
and make it easier to trace requests going through the system.

This removes nginx-full and replaces it with nginx-extras.