Set Request-Id using a boolean flag
I misunderstood how puppet params work. I thought that defining them as
an array in hiera would be fine, but that doesn’t appear to be the case.
I think this is due to how we use vhost_proxies in nodes.pp.

Instead, added a flag with a default value to proxy_vhost and set it for
the vhosts that we care about.

Tested with frontend-app-1 in Vagrant.