Add Naginator plugin
The [naginator
plugin](https://wiki.jenkins-ci.org/display/JENKINS/Naginator+Plugin)
allows us to retry failed builds. Its only intended for use with the
smokey tests which fail due to an interaction with signon that we are
not in control of. To start with we would set this job to retry once. If
there was a failure not due to signon, the build will still fail as
normal - this plugin will only help if the failure is temporary.