Ensure directory has correct permissions
To do this we need to create the mongodb system user, instead of relying
on the mongo package to do it. This is because we need the mount point to be
mounted before installing and starting mongo, and the module we use
doenst provide anyway of creating the directory after the package has
been installed but before the server has been started.