Force domain to be localdomain
This is due to the new template we are using not having a domain set. It
is difficult to force this in the bootstrap script due to the way ubuntu
requires the domain to be set. This should be fine until we want to set
the domain to something other than localdomain, hence the warning.