Re-use Request-Id if there is one
Spotlight makes requests to the public FQDN for backdrop, and tries to
pass the Request-Id that was set by the SSL termination along. We should
proxy that through as well, so that we have a single (hopefully unique)
ID for all requests.

This does mean that we’re trusting content from the internet. We can do
something more if this turns out to be a problem / abused.