Vendor the deprecated elasticsearch module
This module is deprecated, we need to make changes to it. As we are the
only ones using it, it makes sense to vendor it rather than try and
reopen the module in the attic and get a PR merged. Especially as this
should be temporary as we want to switch to using the official
elasticsearch puppet module.

Command to vendor, from a temporary checkout of the repo in the parent
directory of this repo -

git archive d15521e6cb215a809d92692e76a1d0cac111d9a0 | tar -x -C
../pp-puppet/modules/elasticsearch