Elasticsearch to official module and new version
The GDS module we were using has been deprecated and moved into the
attic and we have been meaning to move to it for a while.

The new module has the concept of a central installation with 1 or more
instances that are running on the node. We have configured 1 instance
with the same configuration as our existing elasticsearch setup.

The empty hash for the config param to the ::elasticsearch class is
there to stop it being a string, which blows up puppet.

We have included the apt repo in our manifest vs using manage_repo as
the later caused a dependency cycle in our code.

Templates need to be included when the instances are setup as they will
curl their configuration into the instance and if it is not setup and
running it will break.