Explicitly choose our allowed SSH ciphers
The symmetric portion of the SSH Transport Protocol (as described in RFC
4253 [1]) has security weaknesses that allowed recovery of up to 32 bits
of plaintext from a block of ciphertext that was encrypted with the
Cipher Block Chaining (CBD) method. New Counter mode algorithms (as
described in RFC4344 [2]) were designed that are not vulnerable to these
types of attacks and these algorithms are now recommended for standard
use.

[1] https://tools.ietf.org/html/rfc4253
[2] https://tools.ietf.org/html/rfc4344