Enable SPDY support
SPDY works per domain. This means SPDY can reduce connections and
multiplex requests to a single domain. For instance, if we had a backing
service on a single domain that a frontend service was quite chatty to,
SPDY should be quite beneficial.

http://www.guypo.com/not-as-spdy-as-you-thought/

Do you see where I’m going with this? ;)