Add big_screen app to frontend servers
This is an app that will serve the static HTML/CSS/JS required for
big screen view.

It consists of a location block on the `www` vhost, such that a
request for:

    www.performance.service.gov.uk/big-screen/carers-allowance

will respond with

    /opt/big_screen/current/index.html

Static CSS and JavaScript will be served instead if they exist on
disk because of try_files.

I do not think that what I've written in `big_screen.pp` is good
Puppet and I'd appreciate any thoughts on how to improve it.