Check which is specific to the agent only.
For some reason (memory according to Tim) logstash is going down and checks are not detecting this due to the logstash web process still running. This makes the check more specific to check the agent is up.

If we still want to check for the web we can bump the process count and
have the less specific check or we can create a separate check.