Only check logging rate in production.
Currently logging rate is so low in preview (and most likely staging)
that it is constantly triggering a critical alert in those environments.

This allows us to deploy to staging without constant critical alerts. If
we find we need the log check back however, Tim has suggested adding a
heartbeat.