Alert when almost full boot partition on backup boxes
May be needed on other boxes but for now backup is the one which is
filling up with kernel updates (which we only saw due to side effects).

These numbers may be nuts. I asked a combination of df, wolfram alpha
and the df man pages but it's late on a friday and I'm tired!