Move celery log level config into backend role
environment.yaml is overwritten by pp-deployment, which does not contain
this key. Therefore hiera cannot find a value for the lookup, and writes
the init file with no value for loglevel which causes celery to fail.

We should re evaulate our hiera setup - this is a
confusing way for things to behave.