Capture stderr for stagecraft
The gc debugging we are doing to try and track down issues with
stagecraft 502ing writes to stderr. It seems gunicorn does not redirect
stderr, as I can see the gc messages when running gunicorn locally.
Adding this should push the messages to the upstart log, so we can track
them back