Update to new location of rabbitmq module
This module has moved, and this fails the build as librarian cannot
download it. In addtion, the ref we fix to has disappeared from the new
repo - although it exists on github.com so must have been force pushed.
On examining the change, it appears to only affect rabbitmq versions
less that 3.4 - we run a later version so should be ok without this. I
held off updating the module as this change does not affect rabbitmq -
it is to get the build working again.