Add deploy user's public key
In order to help us migrate Performance Platform MongoDB data across to the
GOV.UK MongoDB stack, we need access from Jenkins running in Performance
Platform.

This does that.

Rather than add this key in to roles.mongo.yaml, it seemed better to add it
here so that we keep everything together, and so that the hash doesn't get
blatted accidentally.