Change stagecraft port
When on the gov.uk dev vm, the stagecraft port needs to change as it
clashes with a gov.uk app. To avoid the app having to support both
versions of vms, it makes sense to make both vms use the same port until
we have migrated all apps to the gov.uk dev vm.